requirements.txt
autopep8==2.0.1
backoff==2.2.1
Deprecated==1.2.13
dnspython==2.3.0
googleapis-common-protos==1.56.2
grpcio==1.51.1
opentelemetry-api==1.15.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-jaeger==1.15.0
opentelemetry-exporter-jaeger-proto-grpc==1.15.0
opentelemetry-exporter-jaeger-thrift==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-instrumentation-pymongo==0.36b0
opentelemetry-instrumentation-wsgi==0.36b0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
opentelemetry-util-http==0.36b0
protobuf==3.20.3
pycodestyle==2.10.0
pymongo==4.3.3
six==1.16.0
thrift==0.16.0
typing_extensions==4.4.0
wrapt==1.14.1
autopep8==2.0.1
bcrypt==4.0.1
black==23.1.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==0.4.3
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pycodestyle==2.10.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
